Make profile runtime a compile time constant
Can now change the time each test will be profiled for via an
environmental variable at compile time.

In an effort to increase reliability of the timer, timing has been
reverted back to clock()

Use raw function pointer inside profile loop to prevent time spent in
std::function::invoke poluting the flat graph

Fixes #44

Also updates c++0x to c++11

+++ Tests/StdWriter/stdReader.cpp
@@ -10,9 +10,9 @@ using namespace std;
 
 //Tests
 template<class Reader>
-long VerifyRead(testLogger& log);
+int VerifyRead(testLogger& log);
 template<class Reader>
-long VerifyReadString(testLogger& log);
+int VerifyReadString(testLogger& log);
 
 string quote = R"(
 There is a computer disease that anybody who works with computers knows about. 
@@ -29,9 +29,9 @@ bool ValidateMatch( const char *str1, const char *str2,
                                       testLogger& log);
 
 template<class Reader>
-long VerifyGet(testLogger& log);
+int VerifyGet(testLogger& log);
 template<class Reader>
-long VerifyRead(testLogger& log);
+int VerifyRead(testLogger& log);
 
 int main(int argc, const char *argv[])
 {
@@ -89,7 +89,7 @@ void Delete ( IFStreamReader* reader ) { remove(reader->Fname().c_str()); delete
 
 
 template<class Reader>
-long VerifyGet(testLogger& log) {
+int VerifyGet(testLogger& log) {
     long ret = 0;
 
 	Reader *reader = Generator();
@@ -116,7 +116,7 @@ long VerifyGet(testLogger& log) {
 }
 
 template<class Reader>
-long VerifyRead(testLogger& log) {
+int VerifyRead(testLogger& log) {
     long ret = 0;
 
 	Reader *reader = Generator();
@@ -149,7 +149,7 @@ long VerifyRead(testLogger& log) {
 }
 
 template<class Reader>
-long VerifyReadString(testLogger& log) {
+int VerifyReadString(testLogger& log) {
     long ret = 0;
 
 	Reader *reader = Generator();
